refac: define styles in a different file
