+++ components/entities/SelectKind/SelectKind.tsx
@@ -2,6 +2,7 @@ import { Kind } from "wollok-ts/dist/model"
 import React  from "react"
 import { ToggleButton, Text } from "react-native-paper"
 import { View  } from "react-native"
+import { stylesheet } from "./styles"
 
 
 export function SelectKind(props: { kind: Kind, setKind: (value: Kind) => void }) {
@@ -12,14 +13,17 @@ export function SelectKind(props: { kind: Kind, setKind: (value: Kind) => void }
     }
 
     return (
-        <ToggleButton.Row style={{ marginVertical: 15, alignSelf: 'center' }}
+        <ToggleButton.Row style={stylesheet.toggleButtonRow}
             onValueChange={(value) => props.setKind(value as Kind)}
             value={props.kind}>
             {entityKinds.map(aKind => {
                 return (
                     <ToggleButton
                         key={aKind.kind}
-                        style={{ width: 100, backgroundColor: toggleButtonColorByKind(aKind.kind) }}
+                        style={[ 
+                            stylesheet.toggleButton,
+                            { backgroundColor: toggleButtonColorByKind(aKind.kind) } 
+                        ]}
                         icon={() => <View><Text>{aKind.description}</Text></View>}
                         value={aKind.kind}
                     />
