+++ components/entities/SelectKind/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+export const stylesheet = StyleSheet.create({
+    toggleButton: {
+        width: 100
+    },
+    toggleButtonRow: {
+        marginVertical: 15,
+        alignSelf: 'center'
+    }
+})
