+++ pages/Entities/Entities.tsx
@@ -1,18 +1,19 @@
 import { useTheme } from "@react-navigation/native";
 import React, { useState } from "react";
-import { EntityComponent } from "../components/definitions/Entity";
+import { EntityComponent } from "../../components/entities/Entity/Entity";
 import { Module } from 'wollok-ts/dist/model'
-import { StyleSheet, View } from "react-native";
+import { View } from "react-native";
 import { FAB } from 'react-native-paper'
-import { NewEntityModal } from "../components/definitions/NewEntityModal";
+import { NewEntityModal } from "../../components/entities/NewEntityModal/NewEntityModal";
 import { ScrollView } from "react-native-gesture-handler";
+import { stylesheet } from "./styles";
 
 
 
 
 
 export function Entities() {
-    const { colors } = useTheme()
+    const styles = stylesheet(useTheme())
     const [entities, setEntities] = useState<Module[]>([]);
     const [modalVisible, setModalVisible] = useState(false);
   
@@ -31,18 +32,7 @@ export function Entities() {
                 {entities.map(ent => <EntityComponent key={ent.name} entity={ent}/>)}
             </ScrollView>
             <NewEntityModal visible={modalVisible} addEntity={addEntity} setVisible={setModalVisible}/>
-            <FAB icon="plus" onPress={fabPressed} style={{ ...styles.fab, backgroundColor: colors.primary }} />
+            <FAB icon="plus" onPress={fabPressed} style={ styles.fab } />
         </View>
     )
 }
-
-
-
-const styles = StyleSheet.create({
-    fab: {
-        position: 'absolute',
-        margin: 30,
-        right: 0,
-        bottom: 0,
-    }
-});
